Fix core frequency in config.txt to repair UART console
The UART console becomes garbled during boot, shortly after the baud rate
change and the trusted firmware boot.

On Raspberry Pi 3, the mini UART clock is linked to the VPU core clock, so
that when the core clock frequency changes, the UART baud rate will also
change (see [1]).

Fix the core frequency to 250 MHz to repair the UART console.

[1]: https://www.raspberrypi.org/documentation/configuration/uart.md

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>