Add notice to assure linux is booting in DT mode.
Somewhere along the line, our ACPI tables have been fixed sufficiently that recent 5.x (5.13 is where I'm seeing it) fedora/linux kernels compiled for both ACPI and DT try to actually boot the rpi3 in ACPI mode. Which lasts just long enough to get to acpi_register_gsi() and crash the machine, but not before the console has been re-enabled. Making it hard to debug without tossing in an earlycon= parameter and using a serial port.

Lets tweak the docs, because its probably a good idea to force it to DT mode anyway.