Add a commented out fixed core frequency in config.txt
The UART console may become garbled during boot, shortly after the baud rate
change and the trusted firmware boot.

On Raspberry Pi 3, the mini UART clock is linked to the VPU core clock, so
that when the core clock frequency changes, the UART baud rate will also
change (see [1]).

This adds a comment to fix the core frequency to 250 MHz to repair the UART
console, which users can uncomment as needed.

Closes #32

[1]: https://www.raspberrypi.org/documentation/configuration/uart.md